+++ experimental/behave/src/main/kotlin/net/corda/behave/node/Node.kt
@@ -5,11 +5,9 @@ import net.corda.behave.database.DatabaseType
 import net.corda.behave.file.LogSource
 import net.corda.behave.file.currentDirectory
 import net.corda.behave.file.stagingRoot
-import net.corda.behave.logging.getLogger
 import net.corda.behave.monitoring.PatternWatch
 import net.corda.behave.node.configuration.*
 import net.corda.behave.process.JarCommand
-import net.corda.behave.seconds
 import net.corda.behave.service.Service
 import net.corda.behave.service.ServiceSettings
 import net.corda.behave.ssh.MonitoringSSHClient
@@ -20,6 +18,8 @@ import net.corda.core.internal.div
 import net.corda.core.internal.exists
 import net.corda.core.messaging.CordaRPCOps
 import net.corda.core.utilities.NetworkHostAndPort
+import net.corda.core.utilities.loggerFor
+import net.corda.core.utilities.seconds
 import org.apache.commons.io.FileUtils
 import java.net.InetAddress
 import java.nio.file.Path
@@ -35,7 +35,7 @@ class Node(
         private val settings: ServiceSettings = ServiceSettings()
 ) {
 
-    private val log = getLogger<Node>()
+    private val log = loggerFor<Node>()
 
     private val runtimeDirectory = rootDirectory / config.name
 
@@ -220,7 +220,7 @@ class Node(
 
     private fun installApps() {
         val version = config.distribution.version
-        val appDirectory = stagingRoot / "deps" / "corda" / version / "apps"
+        val appDirectory = stagingRoot / "corda" / version / "apps"
         if (appDirectory.exists()) {
             val targetAppDirectory = runtimeDirectory / "cordapps"
             FileUtils.copyDirectory(appDirectory.toFile(), targetAppDirectory.toFile())