+++ docs/source/changelog.rst
@@ -7,6 +7,9 @@ release, see :doc:`upgrade-notes`.
 Unreleased
 ==========
 
+* Added program line argument ``on-unknown-config-keys`` to allow specifying behaviour on unknown node configuration property keys.
+  Values are: [FAIL, WARN, IGNORE], default to FAIL if unspecified.
+
 * Fix CORDA-1229. Setter-based serialization was broken with generic types when the property was stored
   as the raw type, List for example.